### 2020-06-25 #2391 @whedon
[PRE REVIEW]: argopy: A Python library for Argo ocean data analysis
**Submitting author:** @gmaze (<a href="http://orcid.org/0000-0001-7231-2095">Guillaume Maze</a>)
**Repository:** <a href="https://github.com/euroargodev/argopy" target ="_blank">https://github.com/euroargodev/argopy</a>
**Version:** v0.1.4
