### 2020-06-25 #2391 @kthyng
Hi @gmaze! Do you have any suggestions for reviewers, either from [this list](https://bit.ly/joss-reviewers) or elsewhere? Please don't use the "@" to mention them here to limit extraneous email.
