+++ .travis.yml
@@ -27,7 +27,7 @@ install:
 script:
   - |
     if [[ $TRAVIS_JOB_NAME == 'tarball' ]]; then
-      pip wheel . -w dist --no-deps && \
-      check-manifest --verbose && \
+      python -m pep517.build --source --binary . --out-dir dist/
+      check-manifest --verbose
       twine check dist/*
     fi
