### 2020-06-25 f2a6de5
More work on the argovis to as data fetcher
Try better internal integration
