+++ argopy/data_fetchers/argovis.py
@@ -8,109 +8,240 @@
 # query functions either return dictionary objects or error messages.
 #
 
-
+import numpy as np
 import pandas as pd
 import xarray as xr
-import requests
-import pdb
+from .proto import ArgoDataFetcherProto
+from abc import abstractmethod
+from argopy.stores import httpstore
+from argopy.options import OPTIONS
+import json
+import getpass
+from argopy.utilities import list_standard_variables
 
-class ArgovisDataFetcher(object):
-    """
-    Manage access to Argo data through Argovis
-    """
+access_points = ['wmo']
+exit_formats = ['xarray']
+dataset_ids = ['phy']  # First is default
 
-    @staticmethod
-    def urlopenjson(url):
-        resp = requests.get(url)
-        # Consider any status other than 2xx an error
-        if not resp.status_code // 100 == 2:
-            return "Error: Unexpected response {}".format(resp)
-        return resp.json()
 
-    @staticmethod
-    def to_dataframe(profiles):
+class ArgovisDataFetcher(ArgoDataFetcherProto):
+    ###
+    # Methods to be customised for a specific Argovis request
+    ###
+    @abstractmethod
+    def init(self):
+        """ Initialisation for a specific fetcher """
+        pass
+
+    @abstractmethod
+    def cname(self):
+        """ Return a unique string defining the request
+
+            Provide this string to populate meta data and titles
+        """
+        pass
+
+    @property
+    def url(self):
+        """ Return the URL used to download data """
+        pass
+
+    ###
+    # Methods that must not change
+    ###
+    def __init__(self,
+                 ds: str = "",
+                 cache: bool = False,
+                 cachedir: str = "",
+                 **kwargs):
+        """ Instantiate an Argovis Argo data loader
+
+            Parameters
+            ----------
+            ds: 'phy' or 'bgc' # @tylertucker202: Do you provide BGC data as well ?? what is the way to select which data (core or bgc) to retrieve ?
+            cache : False
+            cachedir : None
+        """
+        self.fs = httpstore(cache=cache, cachedir=cachedir, timeout=120)
+        self.definition = 'Argovis Argo data fetcher'
+        self.dataset_id = OPTIONS['dataset'] if ds == '' else ds
+        self.server = 'https://argovis.colorado.edu'
+        self.init(**kwargs)
+        self.key_map = { \
+            'date': 'TIME', \
+            'date_qc': 'TIME_QC', \
+            'lat': 'LATITUDE', \
+            'lon': 'LONGITUDE', \
+            'cycle_number': 'CYCLE_NUMBER', \
+            'DATA_MODE': 'DATA_MODE', \
+            'DIRECTION': 'DIRECTION', \
+            'platform_number': 'PLATFORM_NUMBER', \
+            'position_qc': 'POSITION_QC', \
+            'pres': 'PRES', \
+            'temp': 'TEMP', \
+            'psal': 'PSAL', \
+            'index': 'N_POINTS'
+        }
+
+    def __repr__(self):
+        summary = ["<datafetcher '%s'>" % self.definition]
+        summary.append("Domain: %s" % self.cname())
+        return '\n'.join(summary)
+
+    # This one will be delegated to the httpstore later
+    def open_json(self, url):
+        return self.fs.open(url)
+
+    def json2dataframe(self, profiles):
+        """ convert json data to Pandas DataFrame """
+        # Make sure we deal with a list
+        if isinstance(profiles, list):
+            data = profiles
+        else:
+            data = [profiles]
+        # Transform
         rows = []
-        for profile in profiles:
-            keys = [ x for x in profile.keys() if not x in ['measurements', 'bgcMeas']]
[…]
