### 2020-06-25 a0ddfc1
Update to internal design
Better integration, new box access point, better parent class.
some issues remain
