+++ argopy/data_fetchers/argovis.py
@@ -12,15 +12,17 @@
 import numpy as np
 import pandas as pd
 import xarray as xr
-from .proto import ArgoDataFetcherProto
-from abc import abstractmethod
-from argopy.stores import httpstore
-from argopy.options import OPTIONS
 import json
 import getpass
-from argopy.utilities import list_standard_variables
+from .proto import ArgoDataFetcherProto
+from abc import abstractmethod
 
-access_points = ['wmo']
+from argopy.stores import httpstore
+from argopy.options import OPTIONS
+from argopy.utilities import list_standard_variables
+from argopy.errors import DataNotFound
+
+access_points = ['wmo', 'box']
 exit_formats = ['xarray']
 dataset_ids = ['phy']  # First is default
 
@@ -84,6 +86,20 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
             'index': 'N_POINTS'
         }
 
+    def _format(self, x, typ):
+        """ string formating helper """
+        if typ == 'lon':
+            if x < 0:
+                x = 360. + x
+            return ("%05d") % (x * 100.)
+        if typ == 'lat':
+            return ("%05d") % (x * 100.)
+        if typ == 'prs':
+            return ("%05d") % (np.abs(x) * 10.)
+        if typ == 'tim':
+            return pd.to_datetime(x).strftime('%Y%m%d')
+        return str(x)
+
     def __repr__(self):
         summary = ["<datafetcher '%s'>" % self.definition]
         summary.append("Domain: %s" % self.cname())
@@ -106,6 +122,67 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
                 rows.append(row)
         df = pd.DataFrame(rows)
         return df
+
+    def to_dataframe(self):
+        """ """
+        results = []
+        urls = self.url
+        if isinstance(urls, str):
+            urls = [urls]  # Make sure we deal with a list
+        for url in urls:
+            js = self.fs.open_json(url)
+            if isinstance(js, str):
+                continue
+            df = self.json2dataframe(js)
+            df = df.reset_index()
+            df = df.rename(columns=self.key_map)
+            df = df[[value for value in self.key_map.values()]]
+            results.append(df)
+
+        results = [r for r in results if r is not None]  # Only keep non-empty results
+        if len(results) > 0:
+            df = pd.concat(results, ignore_index=True)
+            df.sort_values(by=['TIME', 'PRES'], inplace=True)
+            df = df.set_index(['N_POINTS'])
+            # df['N_POINTS'] = np.arange(0, len(df['N_POINTS']))  # Re-index to avoid duplicate values
+            return df
+        else:
+            raise DataNotFound("CAN'T FETCH ANY DATA !")
+
+    def to_xarray(self):
+        """ Download and return data as xarray Datasets """
+        ds = self.to_dataframe().to_xarray()
+        ds = ds.sortby(['TIME', 'PRES'])  # should already be sorted by date in decending order
+        ds['N_POINTS'] = np.arange(0, len(ds['N_POINTS']))  # Re-index to avoid duplicate values
+
+        # Set coordinates:
+        # ds = ds.set_coords('N_POINTS')
+        coords = ('LATITUDE', 'LONGITUDE', 'TIME', 'N_POINTS')
+        ds = ds.reset_coords()
+        ds['N_POINTS'] = ds['N_POINTS']
+        # Convert all coordinate variable names to upper case
+        for v in ds.data_vars:
+            ds = ds.rename({v: v.upper()})
+        ds = ds.set_coords(coords)
+
[…]
