+++ argopy/data_fetchers/proto.py
@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
 import pandas as pd
+import numpy as np
 
 class ArgoDataFetcherProto(ABC):
     @abstractmethod
