### 2020-06-25 ad5fb7d
Delete avtest.py
