### 2020-06-26 e8b5214
Update to doc page for argovis
also removed argovis and erddap raw dataset global attributes and replace them with argopy ones
