### 2020-06-26 8a9abf7
Update environment.yml
Now install the github/master branch version instead of the last pypi release
