### 2020-06-26 54073db
Update xarray.py
don't raise an error for sdl interp on profiles, instead try to convert point to profile automatically
