+++ argopy/xarray.py
@@ -608,12 +608,11 @@ class ArgoAccessor:
         else:
             raise ValueError('Standard levels must be a list or a numpy array of positive and sorted values')
 
-        if self._type == 'profile':
-            # ds = self._obj.profile2point()
+        if self._type != 'profile':
             raise InvalidDatasetStructure(
                 "Method only available for a collection of profiles")
-        else:
-            ds = self._obj
+
+        ds = self._obj
 
         # Selecting profiles that have a max(pressure) > max(std_lev) to avoid extrapolation in that direction
         # For levels < min(pressure), first level values of the profile are extended to surface.     
