### 2020-06-26 4f85858
Add dashboard for float at the fetcher level
