### 2020-06-26 2b346f8
Add dashboard to tryit nb
