### 2020-06-27 2fec75a
fix error in plot_profilerType
