### 2020-06-29 #2391 @kthyng
@dhruvbalwada are you interested in reviewing this submission to JOSS? We are asking for reviews within about 6 weeks currently. You can see how the process goes here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
