+++ argopy/errors.py
@@ -8,6 +8,10 @@ class DataNotFound(ValueError):
     """ Raise when a data selection return nothing """
     pass
 
+class FtpPathError(ValueError):
+    """ Raise when the ftp path is not appropriate """
+    pass
+
 
 class NetCDF4FileNotFoundError(FileNotFoundError):
     """
