+++ argopy/data_fetchers/localftp_data.py
@@ -41,7 +41,7 @@ import distributed
 
 from .proto import ArgoDataFetcherProto
 from argopy.errors import NetCDF4FileNotFoundError
-from argopy.utilities import list_standard_variables, load_dict, mapp_dict
+from argopy.utilities import list_standard_variables, load_dict, mapp_dict, check_localftp
 from argopy.options import OPTIONS
 from argopy.stores import filestore, indexstore, indexfilter_wmo, indexfilter_box
 from argopy.plotters import open_dashboard
@@ -114,6 +114,7 @@ class LocalFTPArgoDataFetcher(ArgoDataFetcherProto):
         self.definition = 'Local ftp Argo data fetcher'
         self.dataset_id = OPTIONS['dataset'] if ds == '' else ds
         self.local_ftp = OPTIONS['local_ftp'] if local_ftp == '' else local_ftp
+        check_localftp(self.local_ftp, errors='raise')  # Validate local_ftp
         self.init(**kwargs)
 
     def __repr__(self):
