### 2020-07-01 f907388
Update argopy/errors.py
Co-authored-by: Kevin Balem <quai20@users.noreply.github.com>
