+++ argopy/errors.py
@@ -5,7 +5,7 @@ A bunch of custom errors used in argopy.
 """
 
 class DataNotFound(ValueError):
-    """ Raise when a data selection return nothing """
+    """ Raise when a data selection returns nothing """
     pass
 
 class FtpPathError(ValueError):
