### 2020-07-01 fdecbff
Implement API check
