### 2020-07-01 bbaa5d9
Update whats-new.rst
