+++ docs/whats-new.rst
@@ -22,10 +22,11 @@ v0.1.5 (XX July 2020)
 
 **Breaking changes with previous versions**
 
-- The ``local_ftp`` option of the ``localftp`` data source must now points to the folder where the ``dac`` directory is found. This breaks compatibility with rsynced local FTP copy because rsync does not give a ``dac`` folder (e.g. :issue:`33`). A warning is raised to notify users if any of the DAC name is found at the n-1 path level. (:pr:`34`).
+- The ``local_ftp`` option of the ``localftp`` data source must now points to the folder where the ``dac`` directory is found. This breaks compatibility with rsynced local FTP copy because rsync does not give a ``dac`` folder (e.g. :issue:`33`). An instructive error message is raised to notify users if any of the DAC name is found at the n-1 path level. (:pr:`34`).
 
 **Internals**
 
+- Implement a webAPI availability check in unit testing. This allows for more robust ``erddap`` and ``argovis`` tests that are not only based on internet connectivity only. (:commit:`5a46a39a3368431c6652608ee7241888802f334f`).
 
 
 v0.1.4 (24 June 2020)
