### 2020-07-01 bab9ec5
block_size bug in fsspec
