+++ .github/workflows/apistatus.yml
@@ -8,7 +8,7 @@ on:
     - cron:  '*/3 * * * *'
 
 jobs:
-  check:
+  api-check:
 
     runs-on: ubuntu-latest
 
