+++ .github/workflows/apistatus.yml
@@ -23,5 +23,4 @@ jobs:
         python -m pip install --upgrade pip
         pip install git+http://github.com/euroargodev/argopy.git@api-status
     - name: Check API status
-      run: |
-        python -c 'import argopy; argopy.utilities.save_api_status()'
+      run: python -c "import argopy; argopy.utilities.save_api_status()"
