### 2020-07-01 c5efed5
More parallel options
Can switch from thread vs process pools and support for dask client
also add a progress bar
