### 2020-07-02 #2391 @kthyng
@DamienIrving are you interested in reviewing this submission for JOSS? We are asking for reviews in about 6 weeks.
