### 2020-07-02 #2425 @dhruvbalwada
@kthyng @whedon - how do I stop from getting my email inbox bombarded by JOSS? I am getting an email for what seems to be like a review of every single paper on JOSS.

![Screen Shot 2020-07-02 at 4 14 39 PM](https://user-images.githubusercontent.com/18236610/86416565-27092280-bc7f-11ea-882e-dcf214674583.png)
