### 2020-07-02 #2425 @danielskatz
sorry about that - see the instructions in this thread above, specifically: https://github.com/openjournals/joss-reviews/issues/2425#issuecomment-653246380
