### 2020-07-03 9a88d95
Update README.md
Add JOSS paper badge
