### 2020-07-03 7021392
Major internal refactoring
## modules:
Rename fsspec_wrappers.py to filesystems.py

## Unit tests:
This makes it clearer between testing the facade compared to testing each individual fetcher.
