+++ argopy/tests/test_data_fetchers.py
@@ -1,448 +1,0 @@
-#!/bin/env python
-# -*coding: UTF-8 -*-
-#
-# Test data fetchers
-#
-# This is not designed as it should
-# We need to have:
-# - one class to test the facade API
-# - one class to test specific methods of each backends
-#
-# At this point, we are testing real data fetching both through facade and through direct call to backends.
-
-import os
-import numpy as np
-import xarray as xr
-import shutil
-
-import pytest
-import unittest
-from unittest import TestCase
-
-import argopy
-from argopy import DataFetcher as ArgoDataFetcher
-from argopy.errors import InvalidFetcherAccessPoint, InvalidFetcher, ErddapServerError, \
-    CacheFileNotFound, FileSystemHasNoCache, FtpPathError
-from argopy.utilities import list_available_data_src, isconnected, isAPIconnected, erddap_ds_exists
-
-
-argopy.set_options(api_timeout=3*60)  # From Github actions, requests can take a while
-AVAILABLE_SOURCES = list_available_data_src()
-CONNECTED = isconnected()
-CONNECTEDAPI = {src: False for src in AVAILABLE_SOURCES.keys()}
-if CONNECTED:
-    DSEXISTS = erddap_ds_exists(ds="ArgoFloats")
-    DSEXISTS_bgc = erddap_ds_exists(ds="ArgoFloats-bio")
-    DSEXISTS_ref = erddap_ds_exists(ds="ArgoFloats-ref")
-
-    for src in AVAILABLE_SOURCES.keys():
-        try:
-            CONNECTEDAPI[src] = isAPIconnected(src=src, data=True)
-        except InvalidFetcher:
-            pass
-
-else:
-    DSEXISTS = False
-    DSEXISTS_bgc = False
-    DSEXISTS_ref = False
-
-
-def test_invalid_accesspoint():
-    src = list(AVAILABLE_SOURCES.keys())[0]  # Use the first valid data source
-    with pytest.raises(InvalidFetcherAccessPoint):
-        ArgoDataFetcher(src=src).invalid_accesspoint.to_xarray()  # Can't get data if access point not defined first
-    with pytest.raises(InvalidFetcherAccessPoint):
-        ArgoDataFetcher(src=src).to_xarray()  # Can't get data if access point not defined first
-
-
-def test_invalid_fetcher():
-    with pytest.raises(InvalidFetcher):
-        ArgoDataFetcher(src='invalid_fetcher').to_xarray()
-
-
-# @unittest.skipUnless('localftp' in AVAILABLE_SOURCES, "requires localftp data fetcher")
-# def test_unavailable_accesspoint():
-#     with pytest.raises(InvalidFetcherAccessPoint):
-#         ArgoDataFetcher(src='localftp').region([-85., -45., 10., 20., 0., 100.]).to_xarray()
-
-
-class EntryPoints_AllBackends(TestCase):
-    """ Test main API facade for all available fetching backends and default dataset """
-    ftproot = argopy.tutorial.open_dataset('localftp')[0]
-
-    # kwargs_wmo = [{'WMO': 6901929},
-    #               {'WMO': [6901929, 2901623]},
-    #               {'CYC': 1},
-    #               {'CYC': [1, 6]},
-    #               {'WMO': 6901929, 'CYC': 36},
-    #               {'WMO': 6901929, 'CYC': [5, 45]},
-    #               {'WMO': [6901929, 2901623], 'CYC': 2},
-    #               {'WMO': [6901929, 2901623], 'CYC': [2, 23]},
-    #               {}]
-    #
-    # kwargs_box = [{'BOX': [-60, -40, 40., 60.]},
-    #               {'BOX': [-60, -40, 40., 60., '2007-08-01', '2007-09-01']}]
-
-    def setUp(self):
[…]
