### 2020-07-03 3d6c72e
Update filesystems.py
protocol options not passed properly when using caching system !
