### 2020-07-04 b9b0e4e
Parallel download of wmos
