+++ argopy/data_fetchers/erddap_data.py
@@ -79,7 +79,7 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
                  cachedir: str = "",
                  parallel: bool = False,
                  chunks: str = 'auto',
-                 box_maxsize: list = [10, 10, 50],
+                 chunks_maxsize: list = [],
                  parallel_method: str = 'thread',
                  api_timeout: int = 0,
                  **kwargs):
@@ -103,7 +103,7 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
         self.parallel = parallel
         self.parallel_method = parallel_method
         self.chunks = chunks
-        self.box_maxsize = box_maxsize
+        self.chunks_maxsize = chunks_maxsize
 
         self.init(**kwargs)
         self._init_erddapy()
@@ -284,44 +284,37 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
             return self.fs.cachepath(self.url)
 
     @property
-    def url(self, response=None):
+    def url(self):
         """ Return a URL to download the full data request """
         # Replace erddapy get_download_url()
         # We need to replace it to better handle http responses with by-passing the _check_url_response
         # https://github.com/ioos/erddapy/blob/fa1f2c15304938cd0aa132946c22b0427fd61c81/erddapy/erddapy.py#L247
 
-        # Define constraint to select this box of data:
-        self.define_constraints()  # This will affect self.erddap.constraints
-
-        # Define the list of variables to retrieve
-        self.erddap.variables = self._minimal_vlist
-
-        #
-        dataset_id = self.erddap.dataset_id
+        # First part of the URL:
         protocol = self.erddap.protocol
-        variables = self.erddap.variables
-        if not response:
-            response = self.erddap.response
-        constraints = self.erddap.constraints
+        dataset_id = self.erddap.dataset_id
+        response = self.erddap.response
         url = f"{self.erddap.server}/{protocol}/{dataset_id}.{response}?"
-        if variables:
-            variables = ",".join(variables)
-            url += f"{variables}"
 
-        if constraints:
-            _constraints = copy.copy(constraints)
-            for k, v in _constraints.items():
-                if k.startswith("time"):
-                    _constraints.update({k: parse_dates(v)})
-            _constraints = quote_string_constraints(_constraints)
-            _constraints = "".join([f"&{k}{v}" for k, v in _constraints.items()])
+        # Add variables to retrieve:
+        self.erddap.variables = self._minimal_vlist  # Define the list of variables to retrieve
+        variables = self.erddap.variables
+        variables = ",".join(variables)
+        url += f"{variables}"
 
-            url += f"{_constraints}"
+        # Add constraints:
+        self.define_constraints()  # Define constraint to select this box of data (affect self.erddap.constraints)
+        constraints = self.erddap.constraints
+        _constraints = copy.copy(constraints)
+        for k, v in _constraints.items():
+            if k.startswith("time"):
+                _constraints.update({k: parse_dates(v)})
+        _constraints = quote_string_constraints(_constraints)
+        _constraints = "".join([f"&{k}{v}" for k, v in _constraints.items()])
+        url += f"{_constraints}"
 
+        # Last part:
         url += '&distinct()&orderBy("time,pres")'
-        # In erddapy:
-        # url = _distinct(url, **kwargs)
[…]
