### 2020-07-09 cd67299
Update fetchers.py
