### 2020-07-09 5bc0786
Update test_xarray.py
Add tests for teos10
