+++ argopy/tests/test_xarray.py
@@ -100,3 +100,45 @@ class test_interp_std_levels(unittest.TestCase):
             ds.argo.interp_std_levels([-20, 20, 30, 40, 50])
         with pytest.raises(ValueError):
             ds.argo.interp_std_levels(12)
+
+
+@unittest.skipUnless('erddap' in AVAILABLE_SOURCES, "requires erddap data fetcher")
+@unittest.skipUnless(CONNECTED, "erddap requires an internet connection")
+@unittest.skipUnless(DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server")
+class test_teos10(unittest.TestCase):
+
+    @pytest.fixture(autouse=True)
+    def init_data(self):
+        # Fetch real data to test interpolation
+        try:
+            self.ds_pts_standard = ArgoDataFetcher(src='erddap', mode='standard')\
+                .region([-75, -55, 30., 40., 0, 100., '2011-01-01', '2011-01-15'])\
+                .to_xarray()
+            self.ds_pts_expert = ArgoDataFetcher(src='erddap', mode='expert')\
+                .region([-75, -55, 30., 40., 0, 100., '2011-01-01', '2011-01-15'])\
+                .to_xarray()
+        except ErddapServerError:  # Test is passed when something goes wrong because of the erddap server, not our fault !
+            pass
+        except ValueError:  # Catches value error for incorrect standard levels as inputs
+            pass
+
+    def test_teos10_variables_default(self):
+        ds_list = [self.ds_pts_standard, self.ds_pts_expert]
+        for this in ds_list:
+            for format in ['point', 'profile']:
+                that = this
+                if format == 'profile':
+                    that = that.argo.point2profile()
+                that = that.argo.teos10()
+                assert 'SA' in that.variables
+                assert 'CT' in that.variables
+
+    def test_teos10_variables_single(self):
+        ds_list = [self.ds_pts_standard, self.ds_pts_expert]
+        for this in ds_list:
+            for format in ['point', 'profile']:
+                that = this
+                if format == 'profile':
+                    that = that.argo.point2profile()
+                that = that.argo.teos10(['PV'])
+                assert 'PV' in that.variables
