+++ requirements.txt
@@ -8,3 +8,4 @@ dask>=2.9
 toolz>=0.8.2
 erddapy>=0.5
 fsspec>=0.7.1
+gsw>=3.3.1
