### 2020-07-10 2e1e6c8
Ready for merge with new doc and whats new updated
add the inplace option
