### 2020-07-10 18fee5a
Update whats-new.rst
