### 2020-07-10 502bc27
Update whats-new.rst
