### 2020-07-10 66a0b38
Update whats-new.rst
Fix last release date
