### 2020-07-10 890abb2
Add more default var for erddap
