+++ argopy/data_fetchers/erddap_data.py
@@ -65,8 +65,8 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
         pass
 
     @property
-    def urls(self, response=None):
-        """ Return a list of URLs to download data in chunks """
+    def uri(self):
+        """ Return a list of Unique Ressource Identifier (URI) to download data """
         pass
 
     ###
@@ -280,12 +280,12 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
     def cachepath(self):
         """ Return path to cache file for this request """
         if self.parallel:
-            return [self.fs.cachepath(url) for url in self.urls]
+            return [self.fs.cachepath(uri) for uri in self.uri]
         else:
-            return self.fs.cachepath(self.url)
+            return self.fs.cachepath(self.uri)
 
     @property
-    def url(self):
+    def get_url(self):
         """ Return a URL to download the full data request """
         # Replace erddapy get_download_url()
         # We need to replace it to better handle http responses with by-passing the _check_url_response
@@ -321,7 +321,7 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
     @property
     def N_POINTS(self):
         try:
-            with self.fs.open(self.url.replace('.' + self.erddap.response, '.ncHeader')) as of:
+            with self.fs.open(self.get_url.replace('.' + self.erddap.response, '.ncHeader')) as of:
                 ncHeader = of.read().decode("utf-8")
             lines = [line for line in ncHeader.splitlines() if 'row = ' in line][0]
             return int(lines.split('=')[1].split(';')[0])
@@ -333,9 +333,9 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
 
         # Download data
         if not self.parallel:
-            ds = self.fs.open_dataset(self.url)
+            ds = self.fs.open_dataset(self.uri)
         else:
-            ds = self.fs.open_mfdataset(self.urls, method=self.parallel_method)
+            ds = self.fs.open_mfdataset(self.uri, method=self.parallel_method)
 
         ds = ds.rename({'row': 'N_POINTS'})
 
@@ -370,7 +370,7 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
         ds.attrs['Fetched_by'] = getpass.getuser()
         ds.attrs['Fetched_date'] = pd.to_datetime('now').strftime('%Y/%m/%d')
         ds.attrs['Fetched_constraints'] = self.cname()
-        ds.attrs['Fetched_uri'] = self.url
+        ds.attrs['Fetched_uri'] = self.get_url
         ds = ds[np.sort(ds.data_vars)]
 
         #
@@ -424,8 +424,11 @@ class Fetch_wmo(ErddapArgoDataFetcher):
         self.WMO = WMO
         self.CYC = CYC
 
+        chunks_maxsize_default = {'wmo': 5}  # Default chunk size of floats
         if len(self.chunks_maxsize) == 0:
-            self.chunks_maxsize = [2]  # Default chunk size of floats
+            self.chunks_maxsize = chunks_maxsize_default
+        else:
+            self.chunks_maxsize = {**chunks_maxsize_default, **self.chunks_maxsize}
 
         self.definition = "?"
         if self.dataset_id == 'phy':
@@ -457,32 +460,43 @@ class Fetch_wmo(ErddapArgoDataFetcher):
         listname = self.dataset_id + "_" + listname
         return listname
 
-    # @property
-    # def urls(self):
-    #     urls = []
-    #     for wmo in self.WMO:
-    #         urls.append(Fetch_wmo(WMO=wmo, CYC=self.CYC, ds=self.dataset_id).url)
-    #     return urls
-
     @property
-    def urls(self):
+    def uri(self):
[…]
