+++ argopy/utilities.py
@@ -11,6 +11,7 @@ import sys
 import warnings
 import urllib.request
 import json
+import collections
 
 import importlib
 import locale
@@ -19,6 +20,7 @@ import struct
 import subprocess
 
 import xarray as xr
+import pandas as pd
 import numpy as np
 from scipy import interpolate
 
@@ -28,7 +30,7 @@ import shutil
 
 from argopy.options import OPTIONS, set_options
 from argopy.stores import httpstore
-from argopy.errors import FtpPathError, InvalidFetcher
+from argopy.errors import FtpPathError, InvalidFetcher, InvalidFetcherAccessPoint
 
 path2pkl = pkg_resources.resource_filename('argopy', 'assets/')
 
@@ -481,8 +483,6 @@ def open_etopo1(box, res='l'):
 #  From xarrayutils : https://github.com/jbusecke/xarrayutils/blob/master/xarrayutils/vertical_coordinates.py
 #  Direct integration of those 2 functions to minimize dependencies and possibility of tuning them to our needs
 #
-
-
 def linear_interpolation_remap(z, data, z_regridded, z_dim=None, z_regridded_dim="regridded", output_dim="remapped"):
 
     # interpolation called in xarray ufunc
@@ -529,4 +529,257 @@ def linear_interpolation_remap(z, data, z_regridded, z_dim=None, z_regridded_dim
     ).coords[output_dim]
     return remapped
 
-####################
+
+class chunker():
+    # Default maximum chunks size for all possible request parameters
+    default_chunksize = {'box': {'lon': 20,  # degree
+                         'lat': 20,  # degree
+                         'dpt': 500,  # meters/db
+                         'time': 3 * 30},  # Days
+                         'wmo': {
+                         'wmo': 5,    # Nb of floats
+                         'cyc': 100}}  # Nb of cycles
+
+    def __init__(self,
+               request: dict,
+               chunks: str = 'auto',
+               chunksize: dict = {}):
+        """ Create a request chunker
+
+        Allow to easily split an access point request into chunks
+
+        Parameters
+        ----------
+        request: dict
+            Access point request to be chunked. One of the following:
+            {'box': [lon_min, lon_max, lat_min, lat_max, pres_min, pres_max, datim_min, datim_max]}
+            {'box': [lon_min, lon_max, lat_min, lat_max, datim_min, datim_max]}
+            {'wmo': [wmo1, wmo2, ...], 'cyc': [0,1, ...]}
+        chunks: 'auto' or dict
+            Dictionary with the number of chunks to create for each of request parameters.
+        chunksize: dict, optional
+            Dictionary of the maximum chunk size to allow in 'auto' chunking.
+
+        """
+        self.request = request
+
+        default = self.default_chunksize[[k for k in self.request.keys()][0]]
+        if len(chunksize) == 0:  # chunksize = {}
+            chunksize = default
+        else:  # merge with default:
+            chunksize = {**default, **chunksize}
+        self.chunksize = collections.OrderedDict(sorted(chunksize.items()))
+
+        if chunks == 'auto':  # auto for all
+            chunks = {k: 'auto' for k in self.chunksize.keys()}
+        elif len(chunks) == 0:  # chunks = {}, 1 for all
+            chunks = {k: 1 for k in self.request}
+        else:  # merge with default:
+            chunks = {**default, **chunks}
+        self.chunks = collections.OrderedDict(sorted(chunks.items()))
+
+        if 'box' in self.request:
+            if len(self.request['box']) == 8:
+                self.chunker = self._chunker_box3d
+            elif len(self.request['box']) == 6:
+                self.chunker = self._chunker_box2d
[…]
