+++ argopy/data_fetchers/erddap_data.py
@@ -14,13 +14,14 @@ import pandas as pd
 import numpy as np
 import copy
 import warnings
+import distributed
 
 from abc import abstractmethod
 import getpass
 
 from .proto import ArgoDataFetcherProto
 from argopy.options import OPTIONS
-from argopy.utilities import list_standard_variables
+from argopy.utilities import list_standard_variables, chunker
 from argopy.stores import httpstore
 from argopy.plotters import open_dashboard
 
@@ -79,20 +80,23 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
                  cachedir: str = "",
                  parallel: bool = False,
                  chunks: str = 'auto',
-                 chunks_maxsize: list = [],
+                 chunks_maxsize: dict = {},
                  parallel_method: str = 'thread',
                  api_timeout: int = 0,
                  **kwargs):
         """ Instantiate an ERDDAP Argo data loader
 
-            Parameters
-            ----------
-            ds: 'phy' or 'ref' or 'bgc'
-            cache : False
-            cachedir : None
+        Parameters
+        ----------
+        ds: 'phy' or 'ref' or 'bgc'
+        cache : False
+        cachedir : None
 
-            parallel : False
-            chunks : 'auto'
+        parallel : False
+        chunks : str or dict of int, optional
+            Dictionary with keys given by dimension names and values given by chunk sizes.
+            ``chunks={}`` loads data using a single chunk for all requests
+
         """
         timeout = OPTIONS['api_timeout'] if api_timeout == 0 else api_timeout
         self.fs = httpstore(cache=cache, cachedir=cachedir, timeout=timeout)
@@ -100,6 +104,10 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
         self.dataset_id = OPTIONS['dataset'] if ds == '' else ds
         self.server = api_server
 
+        if not isinstance(parallel, bool):
+            # The parallelisation method is passed through 'parallel':
+            if type(parallel) == distributed.client.Client or parallel == 'thread' or parallel == 'process':
+                parallel_method = parallel
         self.parallel = parallel
         self.parallel_method = parallel_method
         self.chunks = chunks
@@ -110,7 +118,10 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
 
     def __repr__(self):
         summary = ["<datafetcher '%s'>" % self.definition]
-        summary.append("Domain: %s" % self.cname())
+        cname = self.cname()
+        if len(cname) > 40:
+            cname = "%s ... %s" % (cname[0:30], cname[-30:])
+        summary.append("Domain: %s" % cname)
         return '\n'.join(summary)
 
     def _add_attributes(self, this):
@@ -424,12 +435,6 @@ class Fetch_wmo(ErddapArgoDataFetcher):
         self.WMO = WMO
         self.CYC = CYC
 
-        chunks_maxsize_default = {'wmo': 5}  # Default chunk size of floats
-        if len(self.chunks_maxsize) == 0:
-            self.chunks_maxsize = chunks_maxsize_default
-        else:
-            self.chunks_maxsize = {**chunks_maxsize_default, **self.chunks_maxsize}
-
         self.definition = "?"
         if self.dataset_id == 'phy':
             self.definition = 'Ifremer erddap Argo data fetcher for floats'
@@ -457,40 +462,16 @@ class Fetch_wmo(ErddapArgoDataFetcher):
                 listname = [listname]
                 [listname.append("CYC%0.4d" % i) for i in self.CYC]
                 listname = "_".join(listname)
-        listname = self.dataset_id + "_" + listname
+        listname = self.dataset_id + ";" + listname
         return listname
 
     @property
     def uri(self):
[…]
