### 2020-07-28 f2425cb
Minor changes to the text of the JOSS paper
