+++ docs/paper.md
@@ -22,27 +22,27 @@ bibliography: paper.bib
 
 # Summary
 
-Argo is a real-time global ocean *in situ* observing system. It provides highly accurate thousands of ocean measurements 
-every day. The Argo dataset has now accumulated more than 2.3 millions of vertical ocean profiles data and accessing it for scientific 
+Argo is a real-time global ocean *in situ* observing system. It provides thousands of highly accurate ocean measurements 
+every day. The Argo dataset has now accumulated more than 2.3 million vertical ocean profiles and accessing it for scientific 
 analysis remains a challenge.
 
 The Argo expert community, focused on delivering a curated dataset of the best scientific quality possible, has never provided 
-its user base with a Python software to easily access and manipulate Argo measurements: the **argopy** software aims 
-to fill this gap. The **argopy** software can be used to easily fetch and manipulate Argo floats measurements. 
-It is dedicated to scientists without knowledge of the Argo data management system but is also designed to accommodate experts 
+its user base with a Python software package to easily access and manipulate Argo measurements: the **argopy** software aims 
+to fill this gap. The **argopy** software can be used to easily fetch and manipulate measurements from Argo floats. 
+It is dedicated to scientists without knowledge of the Argo data management system but is also designed to accommodate expert 
 requirements.
 
 # Introduction
 
-The ocean is a key component of the Earth climate system. It thus needs a continuous real-time monitoring to help scientists 
-better understand its dynamic and predict its evolution. All around the world, oceanographers have manage to join their
+The ocean is a key component of the Earth climate system. It therefore needs continuous real-time monitoring to help scientists 
+better understand its dynamics and to predict its evolution. All around the world, oceanographers have managed to join their
 efforts and set up a [Global Ocean Observing System](https://www.goosocean.org/) among which *Argo* is a key component. 
 
 Argo is a global network of nearly 4000 autonomous probes measuring pressure, temperature and salinity from the surface 
 to 2000m depth every 10 days. The localisation of these probes is nearly random between the $60^o$ parallels ([see live 
-coverage here](http://map.argo-france.fr)). All probes data are collected by satellite in real-time, processed by several 
-data centers and finally merged in a single dataset (collecting more than 2.3 millions of vertical profiles data as of 
-June 2020) made freely available to anyone through a [ftp server](ftp://ftp.ifremer.fr/ifremer/argo) or [monthly zip 
+coverage here](http://map.argo-france.fr)). Data from the probes are collected by satellite in real-time, processed by several 
+data centers, merged in a single dataset (comprising of more than 2.3 million vertical profiles as of 
+June 2020) and made freely available to anyone through an [ftp server](ftp://ftp.ifremer.fr/ifremer/argo) or [monthly zip 
 snapshots](http://dx.doi.org/10.17882/42182).
 
 The Argo international observation array was initiated in 1999 and soon revolutionised our 
@@ -61,7 +61,7 @@ is able to control its buoyancy and thus to move up and down the water column as
 operate the same program, or cycle, illustrated \autoref{fig:argofloat}. After 9 to 10 days of free drift at a parking 
 depth of about 1000m, a typical Argo float dives down to 2000m and then shoals back to the surface while measuring pressure, 
 temperature and salinity. Once it reaches the surface, the float sends by satellite its measurements to a data center 
-where they are processed in real time and made freely available on the web in less than 24h00.
+where they are processed in real time and made freely available on the web in less than 24 hours.
 
 ![Typical 10 days program, cycle, of an Argo float.\label{fig:argofloat}](_static/argofloats_cycle.png)
 
@@ -69,7 +69,7 @@ where they are processed in real time and made freely available on the web in le
 # Why **argopy** ?
 
 For non-experts of the Argo dataset, it is rather complicated to get access to Argo measurements. Even though data are
[…]
