### 2020-07-28 euroargodev/argopy#41 @DamienIrving
Move contributing.rst to root directory
*This comment is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/2425*.

It's common for software repositories to have a `CONTRIBUTING` file in the root directory (along with `LICENSE`, `README`, `CODE_OF_CONDUCT` etc).

You've already got that information in [docs/contributing.rst](https://github.com/euroargodev/argopy/blob/master/docs/contributing.rst), so it might be a good idea to move that file to the root directory. You'll then just need to update [docs/index.rst](https://github.com/euroargodev/argopy/blob/master/docs/index.rst) with something like `.. include:: ../contributing.rst` so that information still appears on your Read the Docs page.
