### 2020-07-28 euroargodev/argopy#42 @DamienIrving
How to get support?
One of the criteria for the JOSS review (https://github.com/openjournals/joss-reviews/issues/2425) reads as follows:

*Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support*

You've covered 1) and 2) in [contributing.rst](https://github.com/euroargodev/argopy/blob/master/docs/contributing.rst), but people might not realise that the gitter logo in your README file is the place to go for support. It might worth explicitly mentioning in the documentation at the Read the Docs site that people can go to https://gitter.im/Argo-floats/argopy for support?
