### 2020-07-28 #2425 @DamienIrving
I've finished my review. My overall comment is that this package is a much needed resource for the oceanographic community and I commend the authors for taking on the task of developing and releasing argopy. I recommend that this paper be published with JOSS after consideration of a few minor comments relating to the following criteria in the checklist:
* **Community guidelines:** It might be a good idea to explicitly state that users can get support at the argopy gitter channel (https://github.com/euroargodev/argopy/issues/42) and to move the `contributing.rst` file to the root directory of the GitHub repo (https://github.com/euroargodev/argopy/issues/41)
* **Quality of writing:** I've made some minor suggested changes to the text of the JOSS paper (https://github.com/euroargodev/argopy/pull/40)
