### 2020-08-02 #2425 @dhruvbalwada
argopy is a software that I have always hoped and wished for, and it will really help fill a very important gap that scientists had to climb through to start using Argo data. There is a lot of great functionality that is part of this release, and there are excellent plans to add more features in the near future. I recommend the publication of this article in JOSS.

I have looked through the minor writing edits suggested above, and second them. I have added some addition suggestions to at https://github.com/DamienIrving/argopy/pull/1 (since I not super familiar with github I was not sure how to create them as a pull request to the main directory, but maybe @DamienIrving can merge them before @gmaze merges Damien's request).

I have also created a couple of minor requests - one issue and one pull request.
