### 2020-08-10 euroargodev/argopy#42 @gmaze
Hi @DamienIrving
thanks for your review
Indeed, we'll make explicit in #30 that Gitter is a place to go for support !
g
