### 2020-08-10 3c3ecb3
Update contributing.rst
