+++ docs/contributing.rst
@@ -10,6 +10,8 @@ Contributing to argopy
   Large parts of this document came from the `Pandas Contributing
   Guide <http://pandas.pydata.org/pandas*docs/stable/contributing.html>`_.
 
+If you seek **support** for your argopy usage: `visit the chat room at gitter <https://gitter.im/Argo-floats/argopy>`_.
+
 Where to start?
 ===============
 
@@ -18,7 +20,7 @@ enhancements, and ideas are welcome.
 
 We will complete this document for guidelines with regard to each of these contributions over time.
 
-If you are brand new to *argopy* or open*source development, we recommend going
+If you are brand new to *argopy* or open source development, we recommend going
 through the `GitHub "issues" tab <https://github.com/euroargodev/argopy/issues>`_
 to find issues that interest you. There are a number of issues listed under
 `Documentation <https://github.com/euroargodev/argopy/issues?q=is%3Aissue+is%3Aopen+label%3Adocumentation>`_
@@ -79,7 +81,9 @@ Contributing to the code base
 Data fetchers
 *************
 
-If you want to add your own data fetcher from a new service, then, keep in mind that:
+Introduction
+------------
+If you want to add your own data fetcher for a new service, then, keep in mind that:
 
 * Data fetchers are responsible for:
 
@@ -88,7 +92,7 @@ If you want to add your own data fetcher from a new service, then, keep in mind 
 
 * Data fetchers must:
 
-    * inherit from the `argopy.data_fetchers.proto.ArgoDataFetcherProto`
+    * inherit from the :class:`argopy.data_fetchers.proto.ArgoDataFetcherProto`
     * provide parameters:
 
             *  ``access_points``, eg: ['wmo', 'box']
@@ -106,6 +110,114 @@ If you want to add your own data fetcher from a new service, then, keep in mind 
 
 It is the responsability of the facade API (:class:`argopy.fetchers.ArgoDataFetcher`) to run
 filters according to user level or requests, not the data fetcher.
+
+Detailled guideline
+-------------------
+
+A new data fetcher must comply with:
+
+Inheritance
+~~~~~~~~~~~
+
+-  [ ] Inherit from the :class:`argopy.data_fetchers.proto.ArgoDataFetcherProto`. This enforces minimal internal design
+   compliance.
+
+Auto-discovery for the facade
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Fetcher properties
+^^^^^^^^^^^^^^^^^^
+
+-  [ ] The new fetcher must come with the ``access_points``,
+   ``exit_formats`` and ``dataset_ids`` properties at the top of the
+   file, e.g.:
+
+   ::
+
+       access_points = ['wmo' ,'box']
+       exit_formats = ['xarray']
+       dataset_ids = ['phy', 'bgc']  # First is default
+
+   Values depend on what the new access point can return and what you want to
+   implement. A good start is with the ``wmo`` access point and the
+   ``phy`` dataset ID. The ``xarray`` data format is the minimum
+   required. These variables are used by the facade
+   to auto-discover the fetcher capabilities. The ``dataset_ids``
+   property is used to determine which variables can be retrieved.
+
+Fetcher access points
+^^^^^^^^^^^^^^^^^^^^^
+
+-  [ ] The new fetcher must come at least with a ``Fetch_wmo`` or
+   ``Fetch_wmo`` class, basically one for each of the ``access_points``
+   listed as properties. More generaly we may have a main class that
+   provides the key functionalities to retrieve data from the source,
+   and then classes for each of the ``access_points`` of your fetcher.
+   This pattern could look like this:
+
+   ::
+
+       class NewDataFetcher(ArgoDataFetcherProto)
+       class Fetch_wmo(NewDataFetcher)
[…]
