### 2020-08-11 euroargodev/argopy#41 @gmaze
hi @DamienIrving
I agree that we must have a CONTRIBUTING guideline at the root level on Github.

But instead of moving the documentation rst file there, I'd rather like to create a CONTRIBUTING.md that explicitly points toward the specific page at https://argopy.readthedocs.io/en/latest/contributing.html
g
