### 2020-08-11 a8df17b
Update contributing.rst
Add link to the gitter chat service for simple question asking and direct usage support.
