### 2020-08-11 515ccdf
Update contributing.rst
Add a Development workflow to the contribution guidelines
