### 2020-08-11 fcb3f32
Update test_xarray.py
