+++ docs/contributing.rst
@@ -5,12 +5,12 @@ Contributing to argopy
 .. contents:: Table of contents:
    :local:
 
+First off, thanks for taking the time to contribute!
+
 .. note::
 
-  Large parts of this document came from the `Pandas Contributing
-  Guide <http://pandas.pydata.org/pandas*docs/stable/contributing.html>`_.
-
-:+1::tada: First off, thanks for taking the time to contribute!:tada::+1:
+  Large parts of this document came from the `Xarray <http://xarray.pydata.org/en/stable/contributing.html>`_
+  and `Pandas <http://pandas.pydata.org/pandas*docs/stable/contributing.html>`_ contributing guides.
 
 If you seek **support** for your argopy usage or if you don't want to read
 this whole thing and just have a question: `visit the chat room at gitter <https://gitter.im/Argo-floats/argopy>`_.
@@ -74,11 +74,121 @@ from others.
 `Click here to open an issue with the specific bug reporting template <https://github.com/euroargodev/argopy/issues/new?template=bug_report.md>`_
 
 
+.. _contributing.documentation:
+
+Contributing to the documentation
+=================================
+
+If you're not the developer type, contributing to the documentation is still of
+huge value. You don't even have to be an expert on *argopy* to do so! In fact,
+there are sections of the docs that are worse off after being written by
+experts. If something in the docs doesn't make sense to you, updating the
+relevant section after you figure it out is a great way to ensure it will help
+the next person.
+
+.. contents:: Documentation:
+   :local:
+
+
+About the *argopy* documentation
+--------------------------------
+
+The documentation is written in **reStructuredText**, which is almost like writing
+in plain English, and built using `Sphinx <http://sphinx-doc.org/>`__. The
+Sphinx Documentation has an excellent `introduction to reST
+<http://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html>`__. Review the Sphinx docs to perform more
+complex changes to the documentation as well.
+
+Some other important things to know about the docs:
+
+- The *argopy* documentation consists of two parts: the docstrings in the code
+  itself and the docs in this folder ``argopy/docs/``.
+
+  The docstrings are meant to provide a clear explanation of the usage of the
+  individual functions, while the documentation in this folder consists of
+  tutorial-like overviews per topic together with some other information
+  (what's new, installation, etc).
+
+- The docstrings follow the **Numpy Docstring Standard**, which is used widely
+  in the Scientific Python community. This standard specifies the format of
+  the different sections of the docstring. See `this document
+  <https://github.com/numpy/numpy/blob/master/doc/HOWTO_DOCUMENT.rst.txt>`_
+  for a detailed explanation, or look at some of the existing functions to
+  extend it in a similar manner.
+
+- The tutorials make heavy use of the `ipython directive
+  <http://matplotlib.org/sampledoc/ipython_directive.html>`_ sphinx extension.
+  This directive lets you put code in the documentation which will be run
+  during the doc build. For example:
+
+  .. code:: rst
+
+      .. ipython:: python
+
+          x = 2
+          x ** 3
+
+  will be rendered as::
+
+      In [1]: x = 2
+
+      In [2]: x ** 3
+      Out[2]: 8
+
+  Almost all code examples in the docs are run (and the output saved) during the
+  doc build. This approach means that code examples will always be up to date,
+  but it does make the doc building a bit more complex.
+
+- Our API documentation in ``docs/api.rst`` houses the auto-generated
+  documentation from the docstrings. For classes, there are a few subtleties
[…]
