+++ docs/requirements.txt
@@ -9,3 +9,4 @@ nbsphinx>=0.4
 sphinx-issues>=1.2
 distributed
 cmocean
+gitpython>=3.1.2
