+++ .github/workflows/pythonpackage.yml
@@ -37,7 +37,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
-        pip install distributed
+        pip install distributed aiohttp
 
 #    - name: Install dependencies with conda
 #      shell: bash -l {0}
