+++ argopy/tests/test_xarray.py
@@ -74,24 +74,32 @@ class test_interp_std_levels(unittest.TestCase):
             pass
 
     def test_interpolation(self):
+        if not hasattr(self, 'ds_pts_standard'):
+            self.init_data()
         # Run with success:
         ds = self.ds_pts_standard.argo.point2profile()
         assert 'PRES_INTERPOLATED' in ds.argo.interp_std_levels(
             [20, 30, 40, 50]).dims
 
     def test_points_error(self):
+        if not hasattr(self, 'ds_pts_standard'):
+            self.init_data()
         # Try to interpolate points, not profiles
         ds = self.ds_pts_standard
         with pytest.raises(InvalidDatasetStructure):
             ds.argo.interp_std_levels([20, 30, 40, 50])
 
     def test_mode_error(self):
+        if not hasattr(self, 'ds_pts_standard'):
+            self.init_data()
         # Try to interpolate expert data
         ds = self.ds_pts_expert.argo.point2profile()
         with pytest.raises(InvalidDatasetStructure):
             ds.argo.interp_std_levels([20, 30, 40, 50]).dims
 
     def test_std_error(self):
+        if not hasattr(self, 'ds_pts_standard'):
+            self.init_data()
         # Try to interpolate on a wrong axis
         ds = self.ds_pts_standard.argo.point2profile()
         with pytest.raises(ValueError):
