### 2020-08-12 #2425 @gmaze
Hi @kthyng
I've taken into account all reviewers comments (thank you @DamienIrving and @dhruvbalwada ) and basically finished a [more complete contribution guidelines on the documentation webpage](https://argopy.readthedocs.io/en/joss-paper/contributing.html). I've also incorporated all edits to the manuscript.
So, I guess I'm done do address reviews and the JOSS-paper branch is now up to date and ready to merge if you accept my answers to reviews.
Do you want me to formally address reviews in the comments above ? I've done it directly in the PR and issues raised on the software repo.
g
