+++ argopy/data_fetchers/localftp_data.py
@@ -41,7 +41,7 @@ import distributed
 
 from .proto import ArgoDataFetcherProto
 from argopy.errors import NetCDF4FileNotFoundError
-from argopy.utilities import list_standard_variables, load_dict, mapp_dict, check_localftp
+from argopy.utilities import list_standard_variables, load_dict, mapp_dict, check_localftp, chunker
 from argopy.options import OPTIONS
 from argopy.stores import filestore, indexstore, indexfilter_wmo, indexfilter_box
 from argopy.plotters import open_dashboard
@@ -67,18 +67,18 @@ class LocalFTPArgoDataFetcher(ArgoDataFetcherProto):
         """ Return a unique string defining the request """
         pass
 
-    @abstractmethod
-    def list_argo_files(self, errors: str = 'raise'):
-        """ Set the internal list of absolute path of all files to load
-        This function must defined the attribute: self._list_of_argo_files with a list of path(s)
-
-        Parameters
-        ----------
-        errors: {'raise','ignore'}, optional
-            If 'raise' (default), raises a NetCDF4FileNotFoundError error if any of the requested
-            files cannot be found. If 'ignore', file not found is skipped when fetching data.
-        """
-        pass
+    # @abstractmethod
+    # def list_argo_files(self, errors: str = 'raise'):
+    #     """ Set the internal list of absolute path of all files to load
+    #     This function must define the attribute: self._list_of_argo_files with a list of path(s)
+    #
+    #     Parameters
+    #     ----------
+    #     errors: {'raise','ignore'}, optional
+    #         If 'raise' (default), raises a NetCDF4FileNotFoundError error if any of the requested
+    #         files cannot be found. If 'ignore', file not found is skipped when fetching data.
+    #     """
+    #     pass
 
     ###
     # Methods that must not change
@@ -89,32 +89,53 @@ class LocalFTPArgoDataFetcher(ArgoDataFetcherProto):
                  cache: bool = False,
                  cachedir: str = "",
                  dimension: str = 'point',
+                 errors: str = 'raise',
+                 parallel: bool = False,
+                 chunks: str = 'auto',
+                 chunks_maxsize: dict = {},
+                 parallel_method: str = 'thread',
                  **kwargs):
         """ Init fetcher
 
-            Parameters
-            ----------
-            local_ftp : str
-                Path to the local directory where the 'dac' folder is located.
-            ds : str
-                Name of the dataset to load. Use the global OPTIONS['dataset'] by default.
-            cache : bool
-                Determine if retrieved data should be cached locally or not, False by default.
-            cachedir : str
-                Absolute path to the cache directory
-            dimension : str
-                Main dimension of the output dataset. This can be "profile" to retrieve a collection of
-                profiles, or "point" (default) to have data as a collection of measurements.
-                This can be used to optimise performances
+        Parameters
+        ----------
+        local_ftp : str
+            Path to the local directory where the 'dac' folder is located.
+        ds : str
+            Name of the dataset to load. Use the global OPTIONS['dataset'] by default.
+        cache : bool
+            Determine if retrieved data should be cached locally or not, False by default.
+        cachedir : str
+            Absolute path to the cache directory
+        dimension : str
+            Main dimension of the output dataset. This can be "profile" to retrieve a collection of
[…]
