+++ argopy/tests/test_fetchers_data_localftp.py
@@ -98,7 +98,7 @@ class Backend(TestCase):
                     pass
                 except Exception:
                     print("ERROR LOCALFTP request:\n",
-                          ArgoDataFetcher(src=self.src, ds=dataset).profile(*arg).fetcher.files)
+                          ArgoDataFetcher(src=self.src, ds=dataset).profile(*arg).fetcher.uri)
                     pass
 
     def __testthis_float(self, dataset):
@@ -112,7 +112,7 @@ class Backend(TestCase):
                     pass
                 except Exception:
                     print("ERROR LOCALFTP request:\n",
-                          ArgoDataFetcher(src=self.src, ds=dataset).float(arg).fetcher.files)
+                          ArgoDataFetcher(src=self.src, ds=dataset).float(arg).fetcher.uri)
                     pass
 
     def __testthis_region(self, dataset):
@@ -126,7 +126,7 @@ class Backend(TestCase):
                     pass
                 except Exception:
                     print("ERROR LOCALFTP request:\n",
-                          ArgoDataFetcher(src=self.src, ds=dataset).region(arg).fetcher.files)
+                          ArgoDataFetcher(src=self.src, ds=dataset).region(arg).fetcher.uri)
                     pass
 
     def __testthis(self, dataset):
