+++ argopy/errors.py
@@ -76,6 +76,20 @@ class InvalidFetcher(ValueError):
     pass
 
 
+class InvalidMethod(ValueError):
+    """
+    Raise when trying to do use an Method not available
+    """
+    pass
+
+
+class InvalidDashboard(ValueError):
+    """
+    Raise this when trying to work with a 3rd party online service to display float information
+    """
+    pass
+
+
 class APIServerError(ValueError):
     """
     Raise this when argopy is disrupted by an error due to the Erddap, not argopy machinery
@@ -93,12 +107,5 @@ class ErddapServerError(APIServerError):
 class ArgovisServerError(APIServerError):
     """
     Raise this when argopy is disrupted by an error due to the Erddap, not argopy machinery
-    """
-    pass
-
-
-class InvalidDashboard(ValueError):
-    """
-    Raise this when trying to work with a 3rd party online service to display float information
     """
     pass
