### 2020-08-20 8488c00
CI tests and more consistent box definition
box was defined every where similarly, get homogeneised for chunking to work
