+++ argopy/utilities.py
@@ -480,6 +480,7 @@ def open_etopo1(box, res='l'):
     da.attrs['URI'] = thisurl
     return da
 
+
 #
 #  From xarrayutils : https://github.com/jbusecke/xarrayutils/blob/master/xarrayutils/vertical_coordinates.py
 #  Direct integration of those 2 functions to minimize dependencies and possibility of tuning them to our needs
@@ -535,12 +536,12 @@ class Chunker():
     """ To chunk fetcher requests """
 
     # Default maximum chunks size for all possible request parameters
-    default_chunksize = {'box': {'lon': 20,  # degree
-                                 'lat': 20,  # degree
-                                 'dpt': 500, # meters/db
+    default_chunksize = {'box': {'lon': 20,   # degree
+                                 'lat': 20,   # degree
+                                 'dpt': 500,  # meters/db
                                  'time': 3 * 30},  # Days
-                         'wmo': {'wmo': 5,    # Nb of floats
-                                 'cyc': 100}} # Nb of cycles
+                         'wmo': {'wmo': 5,     # Nb of floats
+                                 'cyc': 100}}  # Nb of cycles
 
     def __init__(self,
                request: dict,
@@ -554,8 +555,8 @@ class Chunker():
         ----------
         request: dict
             Access point request to be chunked. One of the following:
-            {'box': [lon_min, lon_max, lat_min, lat_max, pres_min, pres_max, datim_min, datim_max]}
-            {'box': [lon_min, lon_max, lat_min, lat_max, datim_min, datim_max]}
+            {'box': [lon_min, lon_max, lat_min, lat_max, dpt_min, dpt_max, time_min, time_max]}
+            {'box': [lon_min, lon_max, lat_min, lat_max, dpt_min, dpt_max]}
             {'wmo': [wmo1, wmo2, ...], 'cyc': [0,1, ...]}
         chunks: 'auto' or dict
             Dictionary with request access point as keys and number of chunks to create as values.
@@ -567,11 +568,25 @@ class Chunker():
         """
         self.request = request
 
+        if 'box' in self.request:
+            if len(self.request['box']) == 8:
+                self.this_chunker = self._chunker_box4d
+            elif len(self.request['box']) == 6:
+                self.this_chunker = self._chunker_box3d
+            else:
+                is_box(self.request['box'])
+        elif 'wmo' in self.request:
+            self.this_chunker = self._chunker_wmo
+        else:
+            raise InvalidFetcherAccessPoint("'%s' not valid access point" % ",".join(self.request.keys()))
+
         default = self.default_chunksize[[k for k in self.request.keys()][0]]
         if len(chunksize) == 0:  # chunksize = {}
             chunksize = default
         else:  # merge with default:
             chunksize = {**default, **chunksize}
+        if not isinstance(chunksize, collections.Mapping):
+            raise ValueError("chunksize must be mappable")
         self.chunksize = collections.OrderedDict(sorted(chunksize.items()))
 
         default = {k: 'auto' for k in self.chunksize.keys()}
@@ -579,20 +594,10 @@ class Chunker():
             chunks = default
         elif len(chunks) == 0:  # chunks = {}, i.e. chunk=1 for all
             chunks = {k: 1 for k in self.request}
+        if not isinstance(chunks, collections.Mapping):
+            raise ValueError("chunks must be 'auto' or mappable")
         chunks = {**default, **chunks}
         self.chunks = collections.OrderedDict(sorted(chunks.items()))
-
-        if 'box' in self.request:
[…]
