### 2020-08-20 8970928
CI test for parallel in erddap
