+++ argopy/tests/test_fetchers_data_erddap.py
@@ -12,18 +12,22 @@ import shutil
 
 import pytest
 import unittest
-from unittest import TestCase
 
 import argopy
 from argopy import DataFetcher as ArgoDataFetcher
-from argopy.errors import InvalidFetcher, ErddapServerError, CacheFileNotFound, FileSystemHasNoCache
-from argopy.utilities import list_available_data_src, isconnected, isAPIconnected, erddap_ds_exists
+from argopy.errors import ErddapServerError, CacheFileNotFound, FileSystemHasNoCache
+from argopy.utilities import (
+    list_available_data_src,
+    isconnected,
+    isAPIconnected,
+    erddap_ds_exists,
+)
 
 
-argopy.set_options(api_timeout=3*60)  # From Github actions, requests can take a while
+argopy.set_options(api_timeout=3 * 60)  # From Github actions, requests can take a while
 AVAILABLE_SOURCES = list_available_data_src()
 CONNECTED = isconnected()
-CONNECTEDAPI = isAPIconnected(src='erddap', data=True)
+CONNECTEDAPI = isAPIconnected(src="erddap", data=True)
 if CONNECTEDAPI:
     DSEXISTS = erddap_ds_exists(ds="ArgoFloats")
     DSEXISTS_bgc = erddap_ds_exists(ds="ArgoFloats-bio")
@@ -42,16 +46,19 @@ def is_list_of_integers(lst):
     return all(isinstance(x, int) for x in lst)
 
 
-@unittest.skipUnless('erddap' in AVAILABLE_SOURCES, "requires erddap data fetcher")
+@unittest.skipUnless("erddap" in AVAILABLE_SOURCES, "requires erddap data fetcher")
 @unittest.skipUnless(CONNECTED, "erddap requires an internet connection")
 @unittest.skipUnless(CONNECTEDAPI, "erddap API is not alive")
 # @unittest.skipUnless(False, "skip")
-class Backend(TestCase):
+class Backend(unittest.TestCase):
     """ Test main API facade for all available dataset and access points of the ERDDAP fetching backend """
-    src = 'erddap'
+
+    src = "erddap"
     testcachedir = os.path.expanduser(os.path.join("~", ".argopytest_tmp"))
 
-    @unittest.skipUnless(DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server")
+    @unittest.skipUnless(
+        DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server"
+    )
     def test_cachepath_notfound(self):
         with argopy.set_options(cachedir=self.testcachedir):
             loader = ArgoDataFetcher(src=self.src, cache=True).profile(6902746, 34)
@@ -59,7 +66,9 @@ class Backend(TestCase):
                 loader.fetcher.cachepath
         shutil.rmtree(self.testcachedir)  # Make sure the cache is left empty
 
-    @unittest.skipUnless(DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server")
+    @unittest.skipUnless(
+        DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server"
+    )
     def test_nocache(self):
         with argopy.set_options(cachedir="dummy"):
             loader = ArgoDataFetcher(src=self.src, cache=False).profile(6902746, 34)
@@ -67,11 +76,15 @@ class Backend(TestCase):
             with pytest.raises(FileSystemHasNoCache):
                 loader.fetcher.cachepath
 
-    @unittest.skipUnless(DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server")
+    @unittest.skipUnless(
+        DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server"
+    )
[…]
