### 2020-08-20 0070f92
All unique ressource properties set to uri and more CI tests
