+++ argopy/tests/test_fetchers_facade_data.py
@@ -1,9 +1,3 @@
-#!/bin/env python
-# -*coding: UTF-8 -*-
-#
-# Test data fetcher facade
-#
-
 import numpy as np
 import xarray as xr
 
