+++ argopy/data_fetchers/argovis_data.py
@@ -15,10 +15,11 @@ import getpass
 from .proto import ArgoDataFetcherProto
 from abc import abstractmethod
 import warnings
+import distributed
 
 from argopy.stores import httpstore
 from argopy.options import OPTIONS
-from argopy.utilities import list_standard_variables, format_oneline, is_box
+from argopy.utilities import list_standard_variables, format_oneline, is_box, Chunker
 from argopy.errors import DataNotFound
 from argopy.plotters import open_dashboard
 
@@ -66,6 +67,10 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
                  ds: str = "",
                  cache: bool = False,
                  cachedir: str = "",
+                 parallel: bool = False,
+                 chunks: str = 'auto',
+                 chunks_maxsize: dict = {},
+                 parallel_method: str = 'thread',
                  api_timeout: int = 0,
                  **kwargs):
         """ Instantiate an Argovis Argo data loader
@@ -81,6 +86,16 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
         self.definition = 'Argovis Argo data fetcher'
         self.dataset_id = OPTIONS['dataset'] if ds == '' else ds
         self.server = api_server
+
+        if not isinstance(parallel, bool):
+            # The parallelisation method is passed through 'parallel':
+            if type(parallel) == distributed.client.Client or parallel == 'thread' or parallel == 'process':
+                parallel_method = parallel
+        self.parallel = parallel
+        self.parallel_method = parallel_method
+        self.chunks = chunks
+        self.chunks_maxsize = chunks_maxsize
+
         self.init(**kwargs)
         self.key_map = {
             'date': 'TIME',
@@ -131,30 +146,28 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
         return df
 
     def to_dataframe(self):
-        """ """
-        results = []
-        urls = self.uri
-        if isinstance(urls, str):
-            urls = [urls]  # Make sure we deal with a list
-        for url in urls:
-            js = self.fs.open_json(url)
-            if isinstance(js, str):
-                continue
-            df = self.json2dataframe(js)
+        """ Load Argo data and return a Pandas dataframe """
+
+        # Download data:
+        if not self.parallel:
+            method = 'sequential'
+        else:
+            method = self.parallel_method
+        df_list = self.fs.open_mfjson(self.uri,
+                                    method=method,
+                                    preprocess=self.json2dataframe,
+                                    progress=1)
+
+        # Merge results (list of dataframe):
+        for i, df in enumerate(df_list):
             df = df.reset_index()
             df = df.rename(columns=self.key_map)
             df = df[[value for value in self.key_map.values() if value in df.columns]]
-            results.append(df)
-
-        results = [r for r in results if r is not None]  # Only keep non-empty results
-        if len(results) > 0:
-            df = pd.concat(results, ignore_index=True)
-            df.sort_values(by=['TIME', 'PRES'], inplace=True)
-            df = df.set_index(['N_POINTS'])
-            # df['N_POINTS'] = np.arange(0, len(df['N_POINTS']))  # Re-index to avoid duplicate values
-            return df
-        else:
-            raise DataNotFound("CAN'T FETCH ANY DATA !")
+            df_list[i] = df
+        df = pd.concat(df_list, ignore_index=True)
+        df.sort_values(by=['TIME', 'PRES'], inplace=True)
+        df = df.set_index(['N_POINTS'])
+        return df
 
     def to_xarray(self):
[…]
