+++ argopy/data_fetchers/argovis_data.py
@@ -20,14 +20,15 @@ import distributed
 from argopy.stores import httpstore
 from argopy.options import OPTIONS
 from argopy.utilities import list_standard_variables, format_oneline, is_box, Chunker
-from argopy.errors import DataNotFound
 from argopy.plotters import open_dashboard
 
-access_points = ['wmo', 'box']
-exit_formats = ['xarray']
-dataset_ids = ['phy']  # First is default
-api_server = 'https://argovis.colorado.edu'  # API root url
-api_server_check = api_server + '/selection/overview'  # URL to check if the API is alive
+access_points = ["wmo", "box"]
+exit_formats = ["xarray"]
+dataset_ids = ["phy"]  # First is default
+api_server = "https://argovis.colorado.edu"  # API root url
+api_server_check = (
+    api_server + "/selection/overview"
+)  # URL to check if the API is alive
 
 
 class ArgovisDataFetcher(ArgoDataFetcherProto):
@@ -63,16 +64,18 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
     ###
     # Methods that must not change
     ###
-    def __init__(self,
-                 ds: str = "",
-                 cache: bool = False,
-                 cachedir: str = "",
-                 parallel: bool = False,
-                 chunks: str = 'auto',
-                 chunks_maxsize: dict = {},
-                 parallel_method: str = 'thread',
-                 api_timeout: int = 0,
-                 **kwargs):
+    def __init__(
+        self,
+        ds: str = "",
+        cache: bool = False,
+        cachedir: str = "",
+        parallel: bool = False,
+        chunks: str = "auto",
+        chunks_maxsize: dict = {},
+        parallel_method: str = "thread",
+        api_timeout: int = 0,
+        **kwargs
+    ):
         """ Instantiate an Argovis Argo data loader
 
             Parameters
@@ -81,15 +84,19 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
             cache : False
             cachedir : None
         """
-        timeout = OPTIONS['api_timeout'] if api_timeout == 0 else api_timeout
+        timeout = OPTIONS["api_timeout"] if api_timeout == 0 else api_timeout
         self.fs = httpstore(cache=cache, cachedir=cachedir, timeout=timeout)
-        self.definition = 'Argovis Argo data fetcher'
-        self.dataset_id = OPTIONS['dataset'] if ds == '' else ds
+        self.definition = "Argovis Argo data fetcher"
+        self.dataset_id = OPTIONS["dataset"] if ds == "" else ds
         self.server = api_server
 
         if not isinstance(parallel, bool):
             # The parallelisation method is passed through 'parallel':
-            if type(parallel) == distributed.client.Client or parallel == 'thread' or parallel == 'process':
+            if (
+                type(parallel) == distributed.client.Client
+                or parallel == "thread"
+                or parallel == "process"
+            ):
                 parallel_method = parallel
         self.parallel = parallel
         self.parallel_method = parallel_method
@@ -98,19 +105,19 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
 
         self.init(**kwargs)
         self.key_map = {
-            'date': 'TIME',
-            'date_qc': 'TIME_QC',
-            'lat': 'LATITUDE',
-            'lon': 'LONGITUDE',
-            'cycle_number': 'CYCLE_NUMBER',
-            'DATA_MODE': 'DATA_MODE',
-            'DIRECTION': 'DIRECTION',
[…]
