### 2020-08-21 a2042c5
Fix CI tests and black formatting
