+++ argopy/tests/test_fetchers_data_argovis.py
@@ -5,26 +5,29 @@ import shutil
 
 import pytest
 import unittest
-from unittest import TestCase
 
 import argopy
 from argopy import DataFetcher as ArgoDataFetcher
-from argopy.errors import InvalidFetcher, ArgovisServerError, CacheFileNotFound, FileSystemHasNoCache
-from argopy.utilities import list_available_data_src, isconnected, isAPIconnected
+from argopy.errors import ArgovisServerError, CacheFileNotFound, FileSystemHasNoCache
+from argopy.utilities import (
+    list_available_data_src,
+    isconnected,
+    isAPIconnected,
+    is_list_of_strings,
+)
 
 
-argopy.set_options(api_timeout=3*60)  # From Github actions, requests can take a while
+argopy.set_options(api_timeout=3 * 60)  # From Github actions, requests can take a while
 AVAILABLE_SOURCES = list_available_data_src()
 CONNECTED = isconnected()
-CONNECTEDAPI = isAPIconnected(src='argovis', data=True)
+CONNECTEDAPI = isAPIconnected(src="argovis", data=True)
 
 
-@unittest.skipUnless('argovis' in AVAILABLE_SOURCES, "requires argovis data fetcher")
+@unittest.skipUnless("argovis" in AVAILABLE_SOURCES, "requires argovis data fetcher")
 @unittest.skipUnless(CONNECTED, "argovis requires an internet connection")
 @unittest.skipUnless(CONNECTEDAPI, "argovis API is not alive")
-class Backend(TestCase):
-    """ Test main API facade for all available dataset and access points of the ERDDAP fetching backend """
-    src = 'argovis'
+class Backend(unittest.TestCase):
+    src = "argovis"
     testcachedir = os.path.expanduser(os.path.join("~", ".argopytest_tmp"))
 
     def test_cachepath_notfound(self):
@@ -45,10 +48,11 @@ class Backend(TestCase):
         with argopy.set_options(cachedir=self.testcachedir):
             try:
                 loader = ArgoDataFetcher(src=self.src, cache=True).float(1901393)
-                # 1st call to load from argovis and init cachedir:
+                # 1st call to load and save to cachedir:
                 ds = loader.to_xarray()
                 # 2nd call to load from cached file:
                 ds = loader.to_xarray()
+                print(loader.fetcher.cachepath)
                 assert isinstance(ds, xr.Dataset)
                 assert isinstance(loader.fetcher.cachepath, str)
                 shutil.rmtree(self.testcachedir)
@@ -63,7 +67,27 @@ class Backend(TestCase):
         with argopy.set_options(cachedir=self.testcachedir):
             loader = ArgoDataFetcher(src=self.src, cache=True).profile(6902746, 34)
             try:
-                # 1st call to load from argovis and save to cachedir:
+                # 1st call to load and save to cachedir:
+                ds = loader.to_xarray()
+                # 2nd call to load from cached file
+                ds = loader.to_xarray()
+                assert isinstance(ds, xr.Dataset)
+                assert isinstance(loader.fetcher.cachepath, str)
+                shutil.rmtree(self.testcachedir)
+            except ArgovisServerError:  # Test is passed when something goes wrong because of the argovis server, not our fault !
+                shutil.rmtree(self.testcachedir)
+                pass
+            except Exception:
+                shutil.rmtree(self.testcachedir)
+                raise
+
+    def test_caching_region(self):
+        with argopy.set_options(cachedir=self.testcachedir):
+            loader = ArgoDataFetcher(src=self.src, cache=True).region(
+                [-40, -30, 30, 40, 0, 100, "2011", "2012"]
+            )
+            try:
+                # 1st call to load and save to cachedir:
[…]
