+++ argopy/utilities.py
@@ -33,22 +33,22 @@ from argopy.options import OPTIONS, set_options
 from argopy.stores import httpstore
 from argopy.errors import FtpPathError, InvalidFetcher, InvalidFetcherAccessPoint
 
-path2pkl = pkg_resources.resource_filename('argopy', 'assets/')
+path2pkl = pkg_resources.resource_filename("argopy", "assets/")
 
 
 def clear_cache():
     """ Delete argopy cache folder """
-    if os.path.exists(OPTIONS['cachedir']):
-        shutil.rmtree(OPTIONS['cachedir'])
+    if os.path.exists(OPTIONS["cachedir"]):
+        shutil.rmtree(OPTIONS["cachedir"])
 
 
 def load_dict(ptype):
-    if ptype == 'profilers':
-        with open(os.path.join(path2pkl, 'dict_profilers.pickle'), 'rb') as f:
+    if ptype == "profilers":
+        with open(os.path.join(path2pkl, "dict_profilers.pickle"), "rb") as f:
             loaded_dict = pickle.load(f)
         return loaded_dict
-    elif ptype == 'institutions':
-        with open(os.path.join(path2pkl, 'dict_institutions.pickle'), 'rb') as f:
+    elif ptype == "institutions":
+        with open(os.path.join(path2pkl, "dict_institutions.pickle"), "rb") as f:
             loaded_dict = pickle.load(f)
         return loaded_dict
     else:
@@ -67,26 +67,38 @@ def list_available_data_src():
     AVAILABLE_SOURCES = {}
     try:
         from .data_fetchers import erddap_data as Erddap_Fetchers
-        AVAILABLE_SOURCES['erddap'] = Erddap_Fetchers
+
+        AVAILABLE_SOURCES["erddap"] = Erddap_Fetchers
     except Exception:
-        warnings.warn("An error occured while loading the ERDDAP data fetcher, "
-                      "it will not be available !\n%s\n%s" % (sys.exc_info()[0], sys.exc_info()[1]))
+        warnings.warn(
+            "An error occured while loading the ERDDAP data fetcher, "
+            "it will not be available !\n%s\n%s"
+            % (sys.exc_info()[0], sys.exc_info()[1])
+        )
         pass
 
     try:
         from .data_fetchers import localftp_data as LocalFTP_Fetchers
-        AVAILABLE_SOURCES['localftp'] = LocalFTP_Fetchers
+
+        AVAILABLE_SOURCES["localftp"] = LocalFTP_Fetchers
     except Exception:
-        warnings.warn("An error occured while loading the local FTP data fetcher, "
-                      "it will not be available !\n%s\n%s" % (sys.exc_info()[0], sys.exc_info()[1]))
+        warnings.warn(
+            "An error occured while loading the local FTP data fetcher, "
+            "it will not be available !\n%s\n%s"
+            % (sys.exc_info()[0], sys.exc_info()[1])
+        )
         pass
 
     try:
         from .data_fetchers import argovis_data as ArgoVis_Fetchers
-        AVAILABLE_SOURCES['argovis'] = ArgoVis_Fetchers
+
+        AVAILABLE_SOURCES["argovis"] = ArgoVis_Fetchers
     except Exception:
-        warnings.warn("An error occured while loading the ArgoVis data fetcher, "
-                      "it will not be available !\n%s\n%s" % (sys.exc_info()[0], sys.exc_info()[1]))
+        warnings.warn(
[…]
