### 2020-08-21 ddc8a39
Ensure all data fetchers returns list of str for uri anc cachepath
Also fix CI tests for that
