### 2020-08-21 2899e79
Use tempfile to manage temporary cache folders in CI tests
