### 2020-08-21 bd1c23f
Update filesystems.py
fix bug for data not being init
