### 2020-08-21 72a6af6
Update and fix documentation
