### 2020-08-21 a696c12
Update whats-new.rst
