### 2020-08-24 cef72b5
More doc for parallel
