+++ argopy/tests/__init__.py
@@ -7,6 +7,10 @@ Test suite for argopy continuous integration
 
 import xarray
 import pandas
+import unittest
+import pytest
 
 print("xarray: %s, %s" % (xarray.__version__, xarray.__file__))
 print("pandas: %s, %s" % (pandas.__version__, pandas.__file__))
+print("unittest: %s, %s" % (unittest.__version__, unittest.__file__))
+print("pytest: %s, %s" % (pytest.__version__, pytest.__file__))
