### 2020-08-25 355a3a8
Improve unit testing
Refactor xarray accessor tests
