### 2020-08-25 dc78ab2
Update erddap_data.py
Make use of new open_mfdataset to retrieve more than 5 WMOs in a sequential list of single WMO fetch.
