### 2020-08-25 7e440f8
Update pop cached items
