+++ argopy/stores/filesystems.py
@@ -94,7 +94,10 @@ class argo_store_proto(ABC):
                 raise CacheFileNotFound("No cached file found in %s for: \n%s" % (self.fs.storage[-1], uri))
 
     def _clear_cache_item(self, uri):
-        """ Open fsspec cache registry (pickle file) and remove entry for uri """
+        """ Open fsspec cache registry (pickle file) and remove entry for uri
+
+            This function will be replaced by the pop_from_cache fsspec method introduced in 0.7.4 in another PR
+        """
         # See the "save_cache()" method in:
         # https://filesystem-spec.readthedocs.io/en/latest/_modules/fsspec/implementations/cached.html#WholeFileCacheFileSystem
         fn = os.path.join(self.fs.storage[-1], "cache")
@@ -120,6 +123,7 @@ class argo_store_proto(ABC):
         if self.cache:
             for uri in self.cache_registry:
                 self._clear_cache_item(uri)
+                # self.fs.pop_from_cache(uri)
 
     @abstractmethod
     def open_dataset(self, *args, **kwargs):
