+++ requirements.txt
@@ -7,6 +7,6 @@ netCDF4>=1.3.1
 dask>=2.9
 toolz>=0.8.2
 erddapy>=0.5
-fsspec>=0.7.1
+fsspec>=0.7.4
 gsw>=3.3.1
 tqdm>=4.46.0
