+++ argopy/tests/test_xarray.py
@@ -47,24 +47,24 @@ def test_point2profile2point(ds_pts):
 @requires_connected_erddap_phy
 class Test_interp_std_levels:
     def test_interpolation(self, ds_pts):
-        # Run with success:
+        """Run with success"""
         ds = ds_pts["standard"].argo.point2profile()
         assert "PRES_INTERPOLATED" in ds.argo.interp_std_levels([20, 30, 40, 50]).dims
 
     def test_points_error(self, ds_pts):
-        # Try to interpolate points, not profiles
+        """Try to interpolate points, not profiles"""
         ds = ds_pts["standard"]
         with pytest.raises(InvalidDatasetStructure):
             ds.argo.interp_std_levels([20, 30, 40, 50])
 
     def test_mode_error(self, ds_pts):
-        # Try to interpolate expert data, not possible !
+        """ Try to interpolate expert data, not yet possible """
         ds = ds_pts["expert"].argo.point2profile()
         with pytest.raises(InvalidDatasetStructure):
             ds.argo.interp_std_levels([20, 30, 40, 50]).dims
 
     def test_std_error(self, ds_pts):
-        # Try to interpolate on a wrong axis
+        """Try to interpolate on a wrong axis"""
         ds = ds_pts["standard"].argo.point2profile()
         with pytest.raises(ValueError):
             ds.argo.interp_std_levels([100, 20, 30, 40, 50])
