### 2020-08-25 aa4eff1
Update test_xarray.py
Try new fixture management
