### 2020-08-25 1cb5cc6
More refactoring in tests
