+++ argopy/tests/test_fetchers_data_erddap.py
@@ -14,9 +14,14 @@ from argopy.utilities import (
     isAPIconnected,
     erddap_ds_exists,
     is_list_of_strings,
-    is_list_of_integers
+    is_list_of_integers,
 )
-
+from . import (
+    requires_connected_erddap,
+    requires_connected_erddap_phy,
+    requires_connected_erddap_bgc,
+    requires_connected_erddap_ref,
+)
 
 argopy.set_options(api_timeout=3 * 60)  # From Github actions, requests can take a while
 AVAILABLE_SOURCES = list_available_data_src()
@@ -32,18 +37,13 @@ else:
     DSEXISTS_ref = False
 
 
-@unittest.skipUnless("erddap" in AVAILABLE_SOURCES, "requires erddap data fetcher")
-@unittest.skipUnless(CONNECTED, "erddap requires an internet connection")
-@unittest.skipUnless(CONNECTEDAPI, "erddap API is not alive")
-# @unittest.skipUnless(False, "skip")
-class Backend(unittest.TestCase):
+@requires_connected_erddap
+class Test_Backend:
     """ Test main API facade for all available dataset and access points of the ERDDAP fetching backend """
 
     src = "erddap"
 
-    @unittest.skipUnless(
-        DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server"
-    )
+    @requires_connected_erddap_phy
     def test_cachepath_notfound(self):
         with tempfile.TemporaryDirectory() as testcachedir:
             with argopy.set_options(cachedir=testcachedir):
@@ -51,9 +51,7 @@ class Backend(unittest.TestCase):
                 with pytest.raises(CacheFileNotFound):
                     loader.fetcher.cachepath
 
-    @unittest.skipUnless(
-        DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server"
-    )
+    @requires_connected_erddap_phy
     def test_nocache(self):
         with argopy.set_options(cachedir="dummy"):
             loader = ArgoDataFetcher(src=self.src, cache=False).profile(6902746, 34)
@@ -61,13 +59,13 @@ class Backend(unittest.TestCase):
             with pytest.raises(FileSystemHasNoCache):
                 loader.fetcher.cachepath
 
-    @unittest.skipUnless(
-        DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server"
-    )
+    @requires_connected_erddap_phy
     def test_clearcache(self):
         with tempfile.TemporaryDirectory() as testcachedir:
             with argopy.set_options(cachedir=testcachedir):
-                loader = ArgoDataFetcher(src=self.src, cache=True).float([1901393, 6902746])
+                loader = ArgoDataFetcher(src=self.src, cache=True).float(
+                    [1901393, 6902746]
+                )
                 try:
                     loader.to_xarray()
                     loader.clear_cache()
@@ -78,14 +76,13 @@ class Backend(unittest.TestCase):
                 except Exception:
                     raise
 
-
-    @unittest.skipUnless(
-        DSEXISTS, "erddap requires a valid core Argo dataset from Ifremer server"
-    )
+    @requires_connected_erddap_phy
     def test_caching_float(self):
         with tempfile.TemporaryDirectory() as testcachedir:
             with argopy.set_options(cachedir=testcachedir):
-                loader = ArgoDataFetcher(src=self.src, cache=True).float([1901393, 6902746])
+                loader = ArgoDataFetcher(src=self.src, cache=True).float(
+                    [1901393, 6902746]
+                )
                 try:
                     # 1st call to load and save to cachedir:
[…]
