### 2020-08-26 9614bbb
Major refactoring of open_mfdataset
Move to individual file systems, limited to threads for http, fix bug in multi-processes for local file system
Improved fetcher init options for parallel
Improved Testing
Improved doc
