+++ argopy/stores/filesystems.py
@@ -1,6 +1,7 @@
 import os
 import io
 import types
+import numpy as np
 import xarray as xr
 import pandas as pd
 import requests
@@ -15,11 +16,41 @@ from IPython.core.display import display, HTML
 import concurrent.futures
 from tqdm import tqdm
 import distributed
+import multiprocessing
 
 from argopy.options import OPTIONS
 from argopy.errors import ErddapServerError, FileSystemHasNoCache, CacheFileNotFound, DataNotFound, \
     APIServerError, InvalidMethod
 from abc import ABC, abstractmethod
+
+
+def new_fs(protocol: str = '', cache: bool = False, cachedir: str = "", **kwargs):
+    """ Create a new fsspec file system
+
+    Parameters
+    ----------
+    protocol: str (optional)
+    cache: bool (optional)
+        Use a filecache system on top of the protocol. Default: False
+    cachedir: str
+        Define path to cache directory.
+    **kwargs: (optional)
+        Other arguments passed to :class:`fsspec.filesystem`
+
+    """
+    if not cache:
+        fs = fsspec.filesystem(protocol, **kwargs)
+        cache_registry = None
+    else:
+        fs = fsspec.filesystem("filecache",
+                               target_protocol=protocol,
+                               target_options={**{'simple_links': True, "block_size": 0}, **kwargs},
+                               cache_storage=cachedir,
+                               expiry_time=86400, cache_check=10)
+        # We use a refresh rate for cache of 1 day,
+        # since this is the update frequency of the Ifremer erddap
+        cache_registry = []  # Will hold uri cached by this store instance
+    return fs, cache_registry
 
 
 class argo_store_proto(ABC):
@@ -34,7 +65,7 @@ class argo_store_proto(ABC):
     def __init__(self,
                  cache: bool = False,
                  cachedir: str = "",
-                 **kw):
+                 **kwargs):
         """ Create a file storage system for Argo data
 
             Parameters
@@ -47,17 +78,11 @@ class argo_store_proto(ABC):
         """
         self.cache = cache
         self.cachedir = OPTIONS['cachedir'] if cachedir == '' else cachedir
-        if not self.cache:
-            self.fs = fsspec.filesystem(self.protocol, **kw)
-        else:
-            self.fs = fsspec.filesystem("filecache",
-                                        target_protocol=self.protocol,
-                                        target_options= {**{'simple_links': True, "block_size": 0}, **kw},
-                                        cache_storage=self.cachedir,
-                                        expiry_time=86400, cache_check=10)
-            # We use a refresh rate for cache of 1 day,
-            # since this is the update frequency of the Ifremer erddap
-            self.cache_registry = []  # Will hold uri cached by this store instance
+        self._filesystem_kwargs = kwargs
+        self.fs, self.cache_registry = new_fs(self.protocol,
+                                              self.cache,
+                                              self.cachedir,
+                                              **self._filesystem_kwargs)
 
     def open(self, path, *args, **kwargs):
         self.register(path)
@@ -134,221 +159,6 @@ class argo_store_proto(ABC):
     def read_csv(self):
         pass
 
-    def open_mfdataset(self,
-                       urls,
-                       concat_dim='row',
-                       max_workers = 100,
-                       method: str = 'thread',
-                       progress: bool = False,
-                       concat: bool = True,
-                       preprocess = None,
-                       *args, **kwargs):
-        """ Open multiple urls as a single xarray dataset.
-
-            This is a version of the ``open_dataset`` method that is able to handle a list of urls/paths
-            sequentially or in parallel.
-
-            Use a Threads Pool by default for parallelization.
-
-            Parameters
-            ----------
-            urls: list(str)
-                List of url/path to open
-            concat_dim: str
[…]
