### 2020-08-27 134209d
Update paper.bib
Fix DOI in biblio
