+++ docs/paper.bib
@@ -2,7 +2,7 @@
 	Author = {Roemmich, Dean and Alford, Matthew H. and Claustre, Herv{\'e} and Johnson, Kenneth and King, Brian and Moum, James and Oke, Peter and Owens, W. Brechner and Pouliquen, Sylvie and Purkey, Sarah and Scanderbeg, Megan and Suga, Toshio and Wijffels, Susan and Zilberman, Nathalie and Bakker, Dorothee and Baringer, Molly and Belbeoch, Mathieu and Bittig, Henry C. and Boss, Emmanuel and Calil, Paulo and Carse, Fiona and Carval, Thierry and Chai, Fei and Conchubhair, Diarmuid {\'O}. and d'Ortenzio, Fabrizio and Dall'Olmo, Giorgio and Desbruyeres, Damien and Fennel, Katja and Fer, Ilker and Ferrari, Raffaele and Forget, Gael and Freeland, Howard and Fujiki, Tetsuichi and Gehlen, Marion and Greenan, Blair and Hallberg, Robert and Hibiya, Toshiyuki and Hosoda, Shigeki and Jayne, Steven and Jochum, Markus and Johnson, Gregory C. and Kang, KiRyong and Kolodziejczyk, Nicolas and K{\"o}rtzinger, Arne and Traon, Pierre-Yves Le and Lenn, Yueng-Djern and Maze, Guillaume and Mork, Kjell Arne and Morris, Tamaryn and Nagai, Takeyoshi and Nash, Jonathan and Garabato, Alberto Naveira and Olsen, Are and Pattabhi, Rama Rao and Prakash, Satya and Riser, Stephen and Schmechtig, Catherine and Schmid, Claudia and Shroyer, Emily and Sterl, Andreas and Sutton, Philip and Talley, Lynne and Tanhua, Toste and Thierry, Virginie and Thomalla, Sandy and Toole, John and Troisi, Ariel and Trull, Thomas W. and Turton, Jon and Velez-Belchi, Pedro Joaquin and Walczowski, Waldemar and Wang, Haili and Wanninkhof, Rik and Waterhouse, Amy F. and Waterman, Stephanie and Watson, Andrew and Wilson, Cara and Wong, Annie P. S. and Xu, Jianping and Yasuda, Ichiro},
 	Journal = {Frontiers in Marine Science},
 	Pages = {439},
-	Title = {On the Future of Argo: A Global, Full-Depth, Multi-Disciplinary Array},
+	Title = {On the Future of {Argo}: A Global, Full-Depth, Multi-Disciplinary Array},
 	Volume = {6},
 	Year = {2019},
     Doi = {10.3389/fmars.2019.00439},
@@ -16,7 +16,7 @@
 	Month = {02},
 	Number = {2},
 	Pages = {145--153},
-	Title = {Fifteen years of ocean observations with the global Argo array},
+	Title = {Fifteen years of ocean observations with the global {Argo} array},
 	Volume = {6},
 	Year = {2016},
     Doi = {10.1038/nclimate2872},
@@ -26,36 +26,22 @@
 @article{wong-2020,
 	Author = {Annie Wong and coauthors},
 	Journal = {Journal of Atmospheric and Oceanic Technology},
-	Title = {Argo 1999-2019: two million temperature-salinity profiles and subsurface velocity observations from a global array of profiling floats},
+	Title = {{Argo} 1999-2019: two million temperature-salinity profiles and subsurface velocity observations from a global array of profiling floats},
 	Volume = {in review},
 	Year = {2020}}
 
 @misc{scoop,
-type = "Dataset",
-year = "2017",
-title = "Scoop-Argo : visual quality control for Argo NetCDF data files",
-journal = "",
-editor = "",
-volume = "",
-number = "",
[…]
