### 2020-08-27 #2425 @gmaze
> @gmaze Just read through your paper. Looks good, but can you edit your references? I see that "Argo" is not coming through capitalized, in particular. You'll need to put {} around "Argo", like "{Argo}", to preserve capitalization. Please look at your references in detail to see if there are other tweaks too.

Done !
