+++ docs/paper.bib
@@ -25,15 +25,16 @@
 
 @article{wong-2020,
 	Author = {Annie Wong and coauthors},
-	Journal = {Journal of Atmospheric and Oceanic Technology},
+	Journal = {Frontiers in Marine Science, section Ocean Observation},
 	Title = {{Argo} 1999-2019: two million temperature-salinity profiles and subsurface velocity observations from a global array of profiling floats},
-	Volume = {in review},
+	Volume = {in press},
+	Doi = {10.3389/fmars.2020.00700}
 	Year = {2020}}
 
 @misc{scoop,
 	type = "Dataset",
 	year = "2017",
-	title = "Scoop-Argo : visual quality control for {Argo} NetCDF data files",
+	title = "Scoop-{Argo} : visual quality control for {Argo} NetCDF data files",
 	journal = "",
 	editor = "",
 	volume = "",
