### 2020-08-27 80b915c
Update paper.bib
fix bug in one ref
