+++ docs/paper.bib
@@ -28,7 +28,7 @@
 	Journal = {Frontiers in Marine Science, section Ocean Observation},
 	Title = {{Argo} 1999-2019: two million temperature-salinity profiles and subsurface velocity observations from a global array of profiling floats},
 	Volume = {in press},
-	Doi = {10.3389/fmars.2020.00700}
+	Doi = {10.3389/fmars.2020.00700},
 	Year = {2020}}
 
 @misc{scoop,
