### 2020-08-27 #2425 @gmaze
@kthyng One of the reference has just been accepted, I have the DOI, but it's not online yet, that's why whedon fails to find the doi.
What do you recommend in this situation ?
