### 2020-08-27 #2425 @kthyng
@gmaze The author list is incorrect for the first reference, and NetCDF should be properly capitalized too.
