### 2020-08-27 75dd7af
Update paper.bib
Fix author formating list
