+++ docs/paper.bib
@@ -34,13 +34,13 @@
 @misc{scoop,
 	type = "Dataset",
 	year = "2017",
-	title = "Scoop-{Argo} : visual quality control for {Argo} NetCDF data files",
+	title = "Scoop-{Argo} : visual quality control for {Argo} {NetCDF} data files",
 	journal = "",
 	editor = "",
 	volume = "",
 	number = "",
 	pages = "",
-	author = "Detoc Jerome, Garo Mickael, Carval Thierry, Thepault Baptiste, Mahoudo Pierre",
+	author = "Detoc, Jerome and Garo, Mickael and Carval, Thierry and Thepault, Baptiste and Mahoudo, Pierre",
 	url = "",
 	organization = "",
 	address = "",
