### 2020-08-27 69c0f84
Update paper.bib
fix capitalization
