### 2020-08-27 533225c
Update paper.bib
fix authors order in scoop ref
