+++ docs/paper.bib
@@ -40,7 +40,7 @@
 	volume = "",
 	number = "",
 	pages = "",
-	author = "Detoc, Jerome and Garo, Mickael and Carval, Thierry and Thepault, Baptiste and Mahoudo, Pierre",
+	author = "Detoc, Jerome and Thepault, Baptiste and Carval, Thierry and Mahoudo, Pierre and Garo, Mickael",
 	url = "",
 	organization = "",
 	address = "",
