### 2020-08-27 dbafc88
Update paper.md
Update authors affiliations
