### 2020-08-27 #2425 @gmaze
@kthyng Thanks for your careful reading ! Refs should now be ok.
