### 2020-08-28 eaa5402
Morning Update
- cname management for data fetchers, now in proto
- open_mfdataset: better warning for failled http requests
- better CI tests
- update erddapy requirements version to 0.6
- fix bug #48
