+++ argopy/data_fetchers/erddap_data.py
@@ -14,7 +14,6 @@ import pandas as pd
 import numpy as np
 import copy
 import warnings
-import distributed
 
 from abc import abstractmethod
 import getpass
@@ -29,12 +28,14 @@ from erddapy import ERDDAP
 from erddapy.utilities import parse_dates, quote_string_constraints
 
 
-access_points = ['wmo', 'box']
-exit_formats = ['xarray']
-dataset_ids = ['phy', 'ref', 'bgc']  # First is default
-api_server = 'https://www.ifremer.fr/erddap'  # API root url
+access_points = ["wmo", "box"]
+exit_formats = ["xarray"]
+dataset_ids = ["phy", "ref", "bgc"]  # First is default
+api_server = "https://www.ifremer.fr/erddap"  # API root url
 # api_server = 'https://erddap.ifremer.fr/erddap'  # API root url
-api_server_check = api_server + '/info/ArgoFloats/index.json'  # URL to check if the API is alive
+api_server_check = (
+    api_server + "/info/ArgoFloats/index.json"
+)  # URL to check if the API is alive
 
 
 class ErddapArgoDataFetcher(ArgoDataFetcherProto):
@@ -48,7 +49,7 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
     # Methods to be customised for a specific erddap request
     ###
     @abstractmethod
-    def init(self):
+    def init(self, *args, **kwargs):
         """ Initialisation for a specific fetcher """
         pass
 
@@ -66,7 +67,7 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
         pass
 
     @property
-    def uri(self):
+    def uri(self) -> list:
         """ Return the list of Unique Ressource Identifier (URI) to download data """
         pass
 
@@ -74,17 +75,19 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
     # Methods that must not change
     ###
 
-    def __init__(self,
-                 ds: str = "",
-                 cache: bool = False,
-                 cachedir: str = "",
-                 parallel: bool = False,
-                 parallel_method: str = 'thread',
-                 progress: bool = False,
-                 chunks: str = 'auto',
-                 chunks_maxsize: dict = {},
-                 api_timeout: int = 0,
-                 **kwargs):
+    def __init__(
+        self,
+        ds: str = "",
+        cache: bool = False,
+        cachedir: str = "",
+        parallel: bool = False,
+        parallel_method: str = "thread",
+        progress: bool = False,
+        chunks: str = "auto",
+        chunks_maxsize: dict = {},
+        api_timeout: int = 0,
+        **kwargs,
+    ):
         """ Instantiate an ERDDAP Argo data fetcher
 
         Parameters
@@ -111,17 +114,20 @@ class ErddapArgoDataFetcher(ArgoDataFetcherProto):
         api_timeout: int (optional)
             Erddap request time out in seconds. Set to OPTIONS['api_timeout'] by default.
         """
-        timeout = OPTIONS['api_timeout'] if api_timeout == 0 else api_timeout
+        timeout = OPTIONS["api_timeout"] if api_timeout == 0 else api_timeout
         self.fs = httpstore(cache=cache, cachedir=cachedir, timeout=timeout)
-        self.definition = 'Ifremer erddap Argo data fetcher'
-        self.dataset_id = OPTIONS['dataset'] if ds == '' else ds
+        self.definition = "Ifremer erddap Argo data fetcher"
+        self.dataset_id = OPTIONS["dataset"] if ds == "" else ds
         self.server = api_server
 
         if not isinstance(parallel, bool):
             parallel_method = parallel
             parallel = True
         if parallel_method not in ["thread"]:
[…]
