### 2020-08-31 #2425 @kthyng
Ok great!

Next, please make a tagged release of the repository in github and then archive it at a place like Zenodo. After that, report the version number and archive DOI in the review thread please.
