+++ docs/whats-new.rst
@@ -3,6 +3,15 @@
 What's New
 ==========
 
+v0.1.6 (31 Aug. 2020)
+---------------------
+
+**Features and front-end API**
+
+- **JOSS paper published**. You can now cite argopy with a clean reference. (:pr:`30`) by `G. Maze <http://www.github.com/gmaze>`_ and `K. Balem <http://www.github.com/quai20>`_.
+
+Maze G. and Balem K. (2020). argopy: A Python library for Argo ocean data analysis. *Journal of Open Source Software*, 5(52), 2425 doi: `10.21105/joss.02425 <10.21105/joss.02425>`_.
+
 v0.1.5 (10 July 2020)
 ---------------------
 
