### 2020-08-31 #2425 @gmaze
ok,

I've merged the PR and specific JOSS-paper branch back to master

I've created a new release [version **v0.1.6**](https://github.com/euroargodev/argopy/releases/tag/v0.1.6) with the paper in the main master branch.

The release has been archived at zenodo under doi: 10.5281/zenodo.4009264
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4009264.svg)](https://doi.org/10.5281/zenodo.4009264)

thanks !
