### 2020-08-31 14c4479
Update whats-new.rst
