### 2020-08-31 #2425 @kthyng
Excellent, got the version.

Can you change the title of your Zenodo archive to match the title of your JOSS paper? Also it looks strange as an archive, usually the files are all shown. For example, following are links to two of the last zenodo archives I've dealt with: https://zenodo.org/record/3996524#.X01br_hKjOQ and https://zenodo.org/record/4009224#.X01bsvhKjOQ.  Is yours different just because of the tar file or something more significant?
