### 2020-09-01 #2425 @gmaze
ok:

1. I updated the zenodo archive with the JOSS paper title
1. To have files more explicit, I uploaded a .zip instead of .tar.gz for zenodo to be able to display files.
But this changed the zenodo DOI, it's now: **10.5281/zenodo.4010160**

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4010160.svg)](https://doi.org/10.5281/zenodo.4010160)

([both files comes from the github tagged release v0.1.6](https://github.com/euroargodev/argopy/releases/tag/v0.1.6))
