### 2020-09-01 #2425 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
